requirements.txt
python-telegram-bot>=21
requests>=2.25
